refractor name of medicalInfo to medicalCard
