+++ app/src/androidTest/java/com/github/h3lp3rs/h3lp/MainActivityTest.kt
@@ -1,17 +1,10 @@
 package com.github.h3lp3rs.h3lp
 
-import android.app.Activity
-import android.app.Instrumentation
-import android.content.Intent
-import androidx.test.espresso.Espresso.closeSoftKeyboard
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.*
-import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.Intents.intended
-import androidx.test.espresso.intent.Intents.intending
 import androidx.test.espresso.intent.matcher.IntentMatchers.*
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 
@@ -33,7 +26,7 @@ class MainActivityTest {
     fun enterNameWorksAndSendsIntent() {
         Intents.init()
         onView(withId(R.id.mainGoButton)).perform(click())
-        intended(allOf(hasComponent(MedicalInfo::class.java.name)))
+        intended(allOf(hasComponent(MedicalCardAcivity::class.java.name)))
         Intents.release()
     }
 }
