+++ app/src/main/java/com/github/h3lp3rs/h3lp/MainActivity.kt
@@ -5,8 +5,6 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import android.widget.EditText
-import androidx.core.widget.doOnTextChanged
-import com.google.android.material.textfield.TextInputLayout
 
 const val EXTRA_MESSAGE = "KEY"
 
@@ -20,7 +18,7 @@ class MainActivity : AppCompatActivity() {
     fun sendMessage(view: View) {
         val editText = findViewById<EditText>(R.id.mainName)
         val message = editText.text.toString()
-        val intent = Intent(this, MedicalInfo::class.java).apply {
+        val intent = Intent(this, MedicalCardAcivity::class.java).apply {
             putExtra(EXTRA_MESSAGE, message)
         }
         startActivity(intent)
